comm: changed crc size to 32bits.

+++ src/comm/comm_master.h
@@ -6,7 +6,7 @@
 
 typedef uint16_t comm_addr_t;
 typedef uint16_t comm_seq_t;
-typedef uint16_t comm_crc_t;
+typedef uint32_t comm_crc_t;
 
 int32_t comm_master_tx(comm_addr_t addr, uint8_t *buf);
 int32_t comm_master_rx(uint8_t *buf);
